…hmarking, we get about 2.7x  maximum performance
boost.
2022-09-24 00:14:12
Running ./zz_bench
Run on (112 X 3100 MHz CPU s)
CPU Caches:
  L1 Data 48K (x56)
  L1 Instruction 32K (x56)
  L2 Unified 1280K (x56)
  L3 Unified 43008K (x2)
Load Average: 113.09, 103.01, 100.32
-----------------------------------------------------------
Benchmark                 Time             CPU   Iterations
-----------------------------------------------------------
BM_Parse32_1           5.41 ns         5.41 ns    127614707
BM_ParseAvx32_1        5.41 ns         5.41 ns    131481251
BM_Parse32_2           7.60 ns         7.60 ns    101486965
BM_ParseAvx32_2        7.21 ns         7.21 ns    100662140
BM_Parse32_3           18.6 ns         18.6 ns     37747465
BM_ParseAvx32_3        15.8 ns         15.8 ns     45671267
BM_Parse32_4           21.3 ns         21.3 ns     32821412
BM_ParseAvx32_4        15.4 ns         15.4 ns     45606688
BM_Parse32_5           24.0 ns         24.0 ns     28883162
BM_ParseAvx32_5        14.2 ns         14.2 ns     48919738
BM_Parse32_6           25.3 ns         25.3 ns     28171374
BM_ParseAvx32_6        13.1 ns         13.1 ns     53483401
BM_Parse32_7           26.9 ns         26.9 ns     25914018
BM_ParseAvx32_7        13.1 ns         13.1 ns     53334350
BM_Parse32_8           29.1 ns         29.1 ns     24179552
BM_ParseAvx32_8        13.1 ns         13.1 ns     53368440
BM_Parse32_9           32.8 ns         32.8 ns     21251833
BM_ParseAvx32_9        13.1 ns         13.1 ns     52993618
BM_Parse32_10          35.1 ns         35.1 ns     19933683
BM_ParseAvx32_10       13.1 ns         13.1 ns     52927471
BM_Parse64_1           5.96 ns         5.96 ns    114083173
BM_ParseAvx64_1        5.98 ns         5.98 ns    117008484
BM_Parse64_2           6.74 ns         6.74 ns    106423611
BM_ParseAvx64_2        6.55 ns         6.55 ns    108907164
BM_Parse64_3           18.7 ns         18.7 ns     37244811
BM_ParseAvx64_3        13.1 ns         13.1 ns     54240934
BM_Parse64_4           21.7 ns         21.7 ns     33925661
BM_ParseAvx64_4        12.8 ns         12.8 ns     53840538
BM_Parse64_5           25.2 ns         25.2 ns     28179568
BM_ParseAvx64_5        12.7 ns         12.7 ns     57454421
BM_Parse64_6           28.4 ns         28.4 ns     24626749
BM_ParseAvx64_6        13.1 ns         13.1 ns     54277091
BM_Parse64_7           32.2 ns         32.2 ns     21875611
BM_ParseAvx64_7        12.9 ns         12.9 ns     55145458
BM_Parse64_8           35.0 ns         35.0 ns     19970139
BM_ParseAvx64_8        12.5 ns         12.5 ns     52959753
BM_Parse64_9           38.3 ns         38.3 ns     18326964
BM_ParseAvx64_9        16.6 ns         16.6 ns     42330051
BM_Parse64_10          41.6 ns         41.6 ns     16794702
BM_ParseAvx64_10       16.4 ns         16.4 ns     42469314

in our object level test. End to End benchmarking for object that all are varint fields ,we can get about 20% performance boost